chore(kudos): add kudos record when adding emoji reaction
