+++ packages/server/graphql/public/typeDefs/AddReactjiToReactablePayload.graphql
@@ -1,0 +1,6 @@
+type AddReactjiToReactableSuccess {
+  reactable: Reactable!
+  addedKudos: Kudos
+}
+
+union AddReactjiToReactablePayload = AddReactjiToReactableSuccess | ErrorPayload
