+++ packages/server/graphql/public/typeDefs/Kudos.graphql
@@ -1,0 +1,24 @@
+"""
+A Kudos
+"""
+type Kudos {
+  """
+  Id
+  """
+  id: Int!
+
+  """
+  User who received kudos
+  """
+  receiverUser: User!
+
+  """
+  emoji name
+  """
+  emoji: String!
+
+  """
+  Whether the viewer belongs to the team
+  """
+  isViewerOnTeam: Boolean!
+}
