+++ packages/server/graphql/public/typeDefs/Organization.graphql
@@ -191,4 +191,5 @@ type OrganizationFeatureFlags {
   singleColumnStandups: Boolean!
   publicTeams: Boolean!
   meetingInception: Boolean!
+  kudos: Boolean!
 }
