+++ packages/server/graphql/public/typeDefs/addReactjiToReactable.graphql
@@ -1,0 +1,31 @@
+extend type Mutation {
+  """
+  Add or remove a reactji from a reactable
+  """
+  addReactjiToReactable(
+    """
+    The id of the reactable
+    """
+    reactableId: ID!
+
+    """
+    the type of the
+    """
+    reactableType: ReactableEnum!
+
+    """
+    the id of the reactji to add
+    """
+    reactji: String!
+
+    """
+    If true, remove the reaction, else add it
+    """
+    isRemove: Boolean
+
+    """
+    The id of the meeting
+    """
+    meetingId: ID!
+  ): AddReactjiToReactablePayload!
+}
